✨ feature: 컬렉션 반환 dto 응답 구조 통일 (#87)
* fix: test- redis embed port 중복 문제 수정 (#82)

* ✨[Feature/review] (#84)

* feature: feelStatus api score 값을 사용하도록 변경

* test: CursorPagingUtilTest

* feature: id String -> Integer

* feature: CursorPagingUtil

* feature: String id -> Integer id로 변경

* feature: String id -> Integer id로 변경

* feature: cursor paging

* feature: String id -> Integer Id

* feature: 로그인 성공시 '/auth/signup'로 리다이렉트 되도록 변경

* feature: 로그인 성공시 '/auth/signup'로 리다이렉트 되도록 변경

* feature: 컬렉션 반환시에도 응답 객체 통일

+++ src/main/java/com/beerair/core/auth/infrastructure/jwt/JwtCrypto.java
@@ -27,12 +27,12 @@ import lombok.experimental.UtilityClass;
 public class JwtCrypto implements AuthTokenCrypto {
     private final SignatureAlgorithm signatureAlgorithm;
     private final Key signatureKey;
-    private final int expiration;
+    private final long expiration;
     private final JwtParser jwtParser;
     private final String tokenPurpose;
 
     @Builder
-    private JwtCrypto(String tokenPurpose, String signatureAlgorithm, String signatureKey, int expiration) {
+    private JwtCrypto(String tokenPurpose, String signatureAlgorithm, String signatureKey, long expiration) {
         this.tokenPurpose = tokenPurpose;
         this.signatureAlgorithm = SignatureAlgorithm.forName(signatureAlgorithm);
         this.signatureKey = new SecretKeySpec(
